+++ docs/tailwind.config.ts
@@ -1,15 +1,12 @@
 import { Config } from 'tailwindcss/types/config';
 
-import { preset } from './theme/preset';
+import { preset } from '@marigold/theme-docs/preset';
 
 const config: Config = {
   content: [
     ...preset.content,
 
     // From docs
-    './theme/root.ts',
-    './theme/*.ts',
-    './theme/**/*.ts',
     './ui/**/*.{js,ts,jsx,tsx,mdx}',
     './app/**/*.{js,ts,jsx,tsx,mdx}',