+++ themes/theme-b2b/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @marigold/theme-b2b
 
+## 27.1.10
+
+### Patch Changes
+
+- Updated dependencies [[`290dc0e`](https://github.com/marigold-ui/marigold/commit/290dc0e8b5b5fc1492d391d8e6156bd849f0b37d)]:
+  - @marigold/components@7.8.1
+  - @marigold/system@7.8.1
+  - @marigold/theme-preset@1.3.15
+
 ## 27.1.9
 
 ### Patch Changes